+++ src/libs/kubernetes_nodes.py
@@ -10,7 +10,8 @@ class KubernetesGetNodes:
     def __init__(self,
                  debug_on=True,
                  logger=None,
-                 k8s_api_instance=None):
+                 k8s_api_instance=None,
+                 cluster_name=None):
 
         self.print_helper = PrintHelper('kubernetes_get_nodes', logger)
         self.print_debug = debug_on
@@ -19,6 +20,7 @@ class KubernetesGetNodes:
                                   f"__init__")
 
         self.api_instance = k8s_api_instance
+        self.cluster_name = cluster_name
 
     @handle_exceptions_method
     def get_node_list(self,
@@ -44,6 +46,7 @@ class KubernetesGetNodes:
                 # api_response = self.api_instance.patch_node(node.metadata.name, body)
                 self.print_helper.info_if(self.print_debug,
                                           f"Cluster Nodes: {node.metadata.name}")
+                node_details['cluster'] = self.cluster_name
                 node_details['context'] = active_context
                 node_details['name'] = node.metadata.name
                 if 'kubernetes.io/role' in node.metadata.labels: