+++ src/libs/kubernetes_replicaset.py
@@ -13,7 +13,8 @@ class KubernetesGetRps:
                  debug_on=True,
                  logger=None,
                  k8s_api_instance=None,
-                 k8s_apps_instance=None):
+                 k8s_apps_instance=None,
+                 cluster_name=None):
 
         self.print_helper = PrintHelper('kubernetes_get_rps', logger)
         self.print_debug = debug_on
@@ -25,6 +26,8 @@ class KubernetesGetRps:
                                                     k8s_api_instance)
         self.api_instance = k8s_api_instance
         self.apps_instance = k8s_apps_instance
+
+        self.cluster_name = cluster_name
 
     @handle_exceptions_method
     def get_replica_set(self,
@@ -68,7 +71,8 @@ class KubernetesGetRps:
                             self.print_helper.info_if(self.print_debug,
                                                       f"Replica sets:{st.metadata.name} "
                                                       f"in {st.metadata.namespace}")
-                            details = {'namespace': st.metadata.namespace,
+                            details = {'cluster': self.cluster_name,
+                                       'namespace': st.metadata.namespace,
                                        'available_replicas': st.status.available_replicas,
                                        'replicas': st.status.replicas,
                                        'ready_replicas': st.status.ready_replicas}